# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.PlaceOrderRequest:
    properties:
      price:
        type: number
      quantity:
        type: number
      side:
        type: string
      symbol:
        type: string
      type:
        type: string
    required:
    - quantity
    - side
    - symbol
    - type
    type: object
...
  response.ErrorInfo:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  response.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.ErrorInfo'
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Cryptocurrency quantitative trading API server
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Crypto Quant API
  version: "1.0"
paths:
...
  /data/candles:
    get:
      description: Get historical candle data within a time range
      parameters:
      - description: Trading symbol (e.g., BTCUSDT)
        in: query
        name: symbol
        required: true
        type: string
      - description: Candle interval (e.g., 1h, 1d)
        in: query
        name: interval
        required: true
        type: string
      - description: Start time (RFC3339 format)
        in: query
        name: start
        required: true
        type: string
      - description: End time (RFC3339 format)
        in: query
        name: end
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get candles
      tags:
      - data
  /data/candles/latest:
    get:
      description: Get the most recent candle for a symbol
      parameters:
      - description: Trading symbol (e.g., BTCUSDT)
        in: query
        name: symbol
        required: true
        type: string
      - description: Candle interval (e.g., 1h, 1d)
        in: query
        name: interval
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get latest candle
      tags:
      - data
  /data/collect:
    post:
      description: Collect historical candle data from Binance
      parameters:
      - description: Trading symbol (e.g., BTCUSDT)
        in: query
        name: symbol
        required: true
        type: string
      - description: Candle interval (e.g., 1h, 1d)
        in: query
        name: interval
        required: true
        type: string
      - description: 'Number of days to collect (default: 30)'
        in: query
        name: days
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Collect historical data
      tags:
      - data
  /data/trades:
    get:
      description: Get all trades for a symbol
      parameters:
      - description: Trading symbol (e.g., BTCUSDT)
        in: query
        name: symbol
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get trade history
      tags:
      - data
  /data/validate:
    get:
      description: Check if historical data is available for the requested period
      parameters:
      - description: Trading symbol (e.g., BTCUSDT)
        in: query
        name: symbol
        required: true
        type: string
      - description: Candle interval (e.g., 1h, 1d)
        in: query
        name: interval
        required: true
        type: string
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start
        required: true
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Validate data availability
      tags:
      - data
  /market/price/{symbol}:
    get:
      description: Get the current price for a symbol
      parameters:
      - description: Trading symbol (e.g., BTCUSDT)
        in: path
        name: symbol
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get current price
      tags:
      - market
  /market/prices:
    get:
      description: Get current prices for multiple symbols
      parameters:
      - description: Comma-separated symbols (e.g., BTCUSDT,ETHUSDT)
        in: query
        name: symbols
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get multiple prices
      tags:
      - market
  /market/stream/{symbol}:
    get:
      description: Stream real-time price updates via Server-Sent Events
      parameters:
      - description: Trading symbol (e.g., BTCUSDT)
        in: path
        name: symbol
        required: true
        type: string
      - description: 'Update interval in seconds (default: 1)'
        in: query
        name: interval
        type: integer
      responses:
        "200":
          description: text/event-stream
          schema:
            type: string
      summary: Stream real-time price (SSE)
      tags:
      - market
  /orders:
    post:
      consumes:
      - application/json
      description: Place a new trading order
      parameters:
      - description: Order details
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/handler.PlaceOrderRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
      summary: Place order
      tags:
      - orders
  /orders/{orderId}:
    delete:
      description: Cancel an existing order
      parameters:
      - description: Order ID
        in: path
        name: orderId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Cancel order
      tags:
      - orders
    get:
      description: Get order details by ID
      parameters:
      - description: Order ID
        in: path
        name: orderId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get order
      tags:
      - orders
...
swagger: "2.0"